+++ BalancedCutFinder.h
@@ -12,7 +12,7 @@
 class BalancedCutFinder {
     public:
     BalancedCutFinder(const Graph& graph, const Tree& tree, float factor);
-    int TwoRespectedBalancedCut();
+    float TwoRespectedBalancedCut();
 
     private:
     const Graph& graph;
@@ -24,7 +24,7 @@ class BalancedCutFinder {
     // dict from i to list of edges (u,v) s.t. e[i] is in u->v, but e[i+1] isn't (or e[i] is and i==n-1)
     std::vector<std::vector<Edge>> tree_edge_to_path_out_edges;
     std::map<int, RangeQuery> weights;
-    int global_weight;
+    float global_weight;
 
     // path p-> how much gap we have between e_i and weights[p]
     std::map<int, int> paths_gaps;
@@ -47,12 +47,12 @@ class BalancedCutFinder {
     void AddOutPath(const Edge& edge, bool negative);
 
     // find min cut (within [n*factor, n-n*factor]) through `edge` and another tree edge
-    int MinCutWithEdge(int edge);
+    float MinCutWithEdge(int edge);
 
     // find the min edge inside the path with index `path` in the segment [min_edge, max_edge].
     // together with another subtree of size `other_subtree`, the subtrees together need to have size
     // within [n*factor, n-n*factor]. If is_in_subtree, calculate the subtree rooted in it; otherwise, the outside subtree
-    int MinOfPath(int path, int min_edge, int max_edge, int other_subtree, bool is_in_subtree, RangeQuery &query);
+    float MinOfPath(int path, int min_edge, int max_edge, int other_subtree, bool is_in_subtree, RangeQuery &query);
 
     // adjust min_tree, max_tree to be in the window size we want to check
     void GetMinMaxEdge(int &min_edge, int &max_edge, int min_subtree, int max_subtree, bool is_in_subtree);
